+++ src/actions/CreateIllustration/utils/downloadTemplate.js
@@ -1,0 +1,35 @@
+import { S3Client, GetObjectCommand } from "@aws-sdk/client-s3";
+import { writeBinaryFile } from "@tauri-apps/api/fs";
+import { join, documentDir } from '@tauri-apps/api/path';
+
+export default async function downloadTemplate() {
+
+  const config = {
+    region: 'us-east-1',
+    credentials: {
+      accessKeyId: import.meta.env.VITE_AWS_KEY_ID,
+      secretAccessKey: import.meta.env.VITE_AWS_SECRET_KEY,
+    },
+  };
+
+  const resource = {
+    Bucket: import.meta.env.VITE_AWS_STAGING_BUCKET,
+    Key: "artisan-test/templates/base.ai",
+  };
+
+  try {
+
+    const docsPath = await documentDir();
+    const templateFile = await join(docsPath, 'Artisan', 'Projects', 'template.ai');
+  
+    const s3Client = new S3Client(config);
+    const command = new GetObjectCommand(resource);
+    const response = await s3Client.send(command);
+  
+    return response;
+
+  } catch (error) {
+    console.log(error);
+  }
+
+}
